+++ src/main/java/com/stocktrading/dataloader1/domain/AppEventPublisher.java
@@ -1,5 +1,7 @@
 package com.stocktrading.dataloader1.domain;
 
+import com.stocktrading.dataloader1.domain.event.StockPriceReceivedEvent;
+
 public interface AppEventPublisher {
 
     void publishStockPriceReceivedAsAppEvent(StockPriceReceivedEvent stockPriceReceivedEvent);
