+++ src/main/java/com/stocktrading/dataloader1/domain/StockPriceService.java
@@ -1,5 +1,6 @@
 package com.stocktrading.dataloader1.domain;
 
+import com.stocktrading.dataloader1.domain.event.StockPriceReceivedEvent;
 import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
