+++ src/main/java/com/stocktrading/dataloader1/domain/event/InstrumentSubscriptionStateChangeType.java
@@ -1,4 +1,4 @@
-package com.stocktrading.dataloader1.domain;
+package com.stocktrading.dataloader1.domain.event;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
