+++ src/main/java/com/stocktrading/dataloader1/domain/event/InstrumentSubscriptionStateChangedEvent.java
@@ -1,4 +1,4 @@
-package com.stocktrading.dataloader1.domain;
+package com.stocktrading.dataloader1.domain.event;
 
 import com.stocktrading.dataloader1.domain.model.FinancialInstrumentModel;
 import lombok.Getter;
